move to variant.py
better functions names, and simplified impl of strip_references from https://github.com/ruediger/Boost-Pretty-Printer/issues/34
